class for serialization and decompose crypto_lib
